backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djongo==1.3.6
pymongo==3.12.3
python-dotenv==1.0.0
python-dateutil>=2.8.2 